+++ package.json
@@ -2,8 +2,11 @@
   "name": "editor-pdf-service",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
+    "prebuild": "tslint -c tslint.json -p tsconfig.json --fix",
+    "build": "tsc",
+    "prestart": "npm run build",
     "start": "node .",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -11,11 +14,18 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/vega": "^3.2.0",
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
-    "vega-cli": "^5.7.2"
+    "vega": "^5.8.1"
   },
   "devDependencies": {
-    "nodemon": "^2.0.1"
+    "@types/body-parser": "^1.17.1",
+    "@types/express": "^4.17.2",
+    "@types/node": "^12.12.14",
+    "nodemon": "^2.0.1",
+    "ts-node": "^8.5.4",
+    "tslint": "^5.20.1",
+    "typescript": "3.6.2"
   }
 }
