+++ src/index.ts
@@ -1,0 +1,13 @@
+import express from "express";
+
+const app = express();
+const port = 8080; // default port to listen
+// define a route handler for the default home page
+app.get( "/", ( req, res ) => {
+    res.send("Vega pdf service");
+} );
+
+// start the Express server
+app.listen( port, () => {
+    console.log( `server started at http://localhost:${ port }` );
+} );
